feat(web): migrate hypernative
