+++ apps/web/src/components/dashboard/FirstSteps/index.tsx
@@ -25,8 +25,7 @@ import CheckCircleOutlineRoundedIcon from '@mui/icons-material/CheckCircleOutlin
 import LightbulbOutlinedIcon from '@mui/icons-material/LightbulbOutlined'
 import css from './styles.module.css'
 import { getExplorerLink } from '@safe-global/utils/utils/gateway'
-import { HnDashboardBannerWithNoBalanceCheck } from '@/features/hypernative/components/HnDashboardBanner'
-import { BannerType, useBannerVisibility } from '@/features/hypernative/hooks'
+import { HnDashboardBannerWithNoBalanceCheck, BannerType, useBannerVisibility } from '@/features/hypernative'
 
 const calculateProgress = (items: boolean[]) => {
   const totalNumberOfItems = items.length
