+++ apps/web/src/components/transactions/TxDetails/index.tsx
@@ -44,10 +44,12 @@ import { FEATURES } from '@safe-global/utils/utils/chains'
 import { sameAddress } from '@safe-global/utils/utils/addresses'
 import DecodedData from './TxData/DecodedData'
 import { QueuedTxSimulation } from '../QueuedTxSimulation'
-import { HnQueueAssessmentBanner } from '@/features/hypernative/components/HnQueueAssessmentBanner'
-import { useQueueAssessment } from '@/features/hypernative/hooks/useQueueAssessment'
-import { useShowHypernativeAssessment } from '@/features/hypernative/hooks/useShowHypernativeAssessment'
-import { useHypernativeOAuth } from '@/features/hypernative/hooks/useHypernativeOAuth'
+import {
+  HnQueueAssessmentBanner,
+  useQueueAssessment,
+  useShowHypernativeAssessment,
+  useHypernativeOAuth,
+} from '@/features/hypernative'
 
 export const NOT_AVAILABLE = 'n/a'
