+++ apps/web/src/components/transactions/TxSummary/index.tsx
@@ -20,10 +20,12 @@ import { useHasFeature } from '@/hooks/useChains'
 import TxStatusLabel from '@/components/transactions/TxStatusLabel'
 import { FEATURES } from '@safe-global/utils/utils/chains'
 import { ellipsis } from '@safe-global/utils/utils/formatters'
-import { HnQueueAssessment } from '@/features/hypernative/components/HnQueueAssessment'
-import { useQueueAssessment } from '@/features/hypernative/hooks/useQueueAssessment'
-import { useShowHypernativeAssessment } from '@/features/hypernative/hooks/useShowHypernativeAssessment'
-import { useHypernativeOAuth } from '@/features/hypernative/hooks/useHypernativeOAuth'
+import {
+  HnQueueAssessment,
+  useQueueAssessment,
+  useShowHypernativeAssessment,
+  useHypernativeOAuth,
+} from '@/features/hypernative'
 import { getSafeTxHashFromTxId } from '@/utils/transactions'
 
 type TxSummaryProps = {
