+++ apps/web/src/features/hypernative/components/HnFeature/HnFeature.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@/tests/test-utils'
 import HnFeature from './HnFeature'
 import { BannerType } from '../../hooks/useBannerStorage'
-import * as useIsHypernativeFeatureHook from '../../hooks/useIsHypernativeFeature'
+import * as useIsHypernativeEnabledHook from '../../hooks/useIsHypernativeEnabled'
 import * as useBannerVisibilityHook from '../../hooks/useBannerVisibility'
 
 describe('HnFeature', () => {
@@ -13,7 +13,7 @@ describe('HnFeature', () => {
 
   describe('when feature is not enabled', () => {
     it('should not render children', () => {
-      jest.spyOn(useIsHypernativeFeatureHook, 'useIsHypernativeFeature').mockReturnValue(false)
+      jest.spyOn(useIsHypernativeEnabledHook, 'useIsHypernativeEnabled').mockReturnValue(false)
       jest.spyOn(useBannerVisibilityHook, 'useBannerVisibility').mockReturnValue({
         showBanner: true,
         loading: false,
@@ -32,7 +32,7 @@ describe('HnFeature', () => {
 
   describe('when feature is enabled but banner should not show', () => {
     it('should not render children when showBanner is false', () => {
-      jest.spyOn(useIsHypernativeFeatureHook, 'useIsHypernativeFeature').mockReturnValue(true)
+      jest.spyOn(useIsHypernativeEnabledHook, 'useIsHypernativeEnabled').mockReturnValue(true)
       jest.spyOn(useBannerVisibilityHook, 'useBannerVisibility').mockReturnValue({
         showBanner: false,
         loading: false,
@@ -49,7 +49,7 @@ describe('HnFeature', () => {
     })
 
     it('should not render children when loading is true', () => {
-      jest.spyOn(useIsHypernativeFeatureHook, 'useIsHypernativeFeature').mockReturnValue(true)
+      jest.spyOn(useIsHypernativeEnabledHook, 'useIsHypernativeEnabled').mockReturnValue(true)
       jest.spyOn(useBannerVisibilityHook, 'useBannerVisibility').mockReturnValue({
         showBanner: true,
         loading: true,
@@ -68,7 +68,7 @@ describe('HnFeature', () => {
 
   describe('when feature is enabled and banner should show', () => {
     it('should render children when all conditions are met', () => {
-      jest.spyOn(useIsHypernativeFeatureHook, 'useIsHypernativeFeature').mockReturnValue(true)
+      jest.spyOn(useIsHypernativeEnabledHook, 'useIsHypernativeEnabled').mockReturnValue(true)
       jest.spyOn(useBannerVisibilityHook, 'useBannerVisibility').mockReturnValue({
         showBanner: true,
         loading: false,
@@ -84,7 +84,7 @@ describe('HnFeature', () => {
     })
 
     it('should work with BannerType.Pending', () => {
-      jest.spyOn(useIsHypernativeFeatureHook, 'useIsHypernativeFeature').mockReturnValue(true)
+      jest.spyOn(useIsHypernativeEnabledHook, 'useIsHypernativeEnabled').mockReturnValue(true)
       jest.spyOn(useBannerVisibilityHook, 'useBannerVisibility').mockReturnValue({
         showBanner: true,
         loading: false,
@@ -101,7 +101,7 @@ describe('HnFeature', () => {
     })
 
     it('should work with BannerType.Promo', () => {
-      jest.spyOn(useIsHypernativeFeatureHook, 'useIsHypernativeFeature').mockReturnValue(true)
+      jest.spyOn(useIsHypernativeEnabledHook, 'useIsHypernativeEnabled').mockReturnValue(true)
       jest.spyOn(useBannerVisibilityHook, 'useBannerVisibility').mockReturnValue({
         showBanner: true,
         loading: false,
