+++ apps/web/src/features/hypernative/components/HnQueueAssessmentBanner/HnQueueAssessmentBanner.tsx
@@ -5,9 +5,9 @@ import type { ThreatAnalysisResults } from '@safe-global/utils/features/safe-shi
 import type { AsyncResult } from '@safe-global/utils/hooks/useAsync'
 import ExternalLink from '@/components/common/ExternalLink'
 import { Severity } from '@safe-global/utils/features/safe-shield/types'
-import { useHypernativeOAuth } from '@/features/hypernative/hooks/useHypernativeOAuth'
-import { useAssessmentUrl } from '@/features/hypernative/hooks/useAssessmentUrl'
-import { useHnAssessmentSeverity } from '@/features/hypernative/hooks/useHnAssessmentSeverity'
+import { useHypernativeOAuth } from '../../hooks/useHypernativeOAuth'
+import { useAssessmentUrl } from '../../hooks/useAssessmentUrl'
+import { useHnAssessmentSeverity } from '../../hooks/useHnAssessmentSeverity'
 import LockIcon from '@/public/images/common/lock-small.svg'
 
 interface HnQueueAssessmentBannerProps {
