+++ apps/web/src/features/hypernative/hooks/useIsHypernativeEligible.ts
@@ -1,6 +1,6 @@
 import { useIsOutreachSafe } from '@/features/targetedFeatures/hooks/useIsOutreachSafe'
-import { useIsHypernativeGuard } from '@/features/hypernative/hooks/useIsHypernativeGuard'
-import { HYPERNATIVE_ALLOWLIST_OUTREACH_ID } from '@/features/hypernative/constants'
+import { useIsHypernativeGuard } from './useIsHypernativeGuard'
+import { HYPERNATIVE_ALLOWLIST_OUTREACH_ID } from '../constants'
 
 export type HypernativeEligibility = {
   isHypernativeEligible: boolean
