+++ apps/web/src/features/hypernative/hooks/useIsHypernativeEnabled.ts
@@ -7,7 +7,7 @@ import { FEATURES } from '@safe-global/utils/utils/chains'
  *
  * @returns true if Hypernative features are enabled, false otherwise
  */
-export const useIsHypernativeFeature = (): boolean => {
+export const useIsHypernativeEnabled = (): boolean => {
   const hasFeature = useHasFeature(FEATURES.HYPERNATIVE)
   // Return false if feature is undefined or false
   return hasFeature === true
