+++ apps/web/src/features/safe-shield/components/HypernativeInfo/index.tsx
@@ -3,8 +3,8 @@ import { Box, Button, SvgIcon, Stack, Typography } from '@mui/material'
 import OpenInNewRoundedIcon from '@mui/icons-material/OpenInNewRounded'
 import SafeShieldLogo from '@/public/images/safe-shield/safe-shield-logo-no-text.svg'
 import InfoIcon from '@/public/images/notifications/info.svg'
-import { HypernativeTooltip } from '@/features/hypernative/components/HypernativeTooltip'
-import type { HypernativeAuthStatus } from '@/features/hypernative/hooks/useHypernativeOAuth'
+import { HypernativeTooltip } from '@/features/hypernative'
+import type { HypernativeAuthStatus } from '@/features/hypernative'
 
 export const HypernativeInfo = ({
   hypernativeAuth,
