+++ apps/web/src/features/safe-shield/hooks/__tests__/useNestedThreatAnalysis.test.tsx
@@ -2,7 +2,7 @@ import { renderHook, waitFor } from '@testing-library/react'
 import { useNestedThreatAnalysis } from '../useNestedThreatAnalysis'
 import { Severity, StatusGroup, ThreatStatus } from '@safe-global/utils/features/safe-shield/types'
 import type { SafeTransaction } from '@safe-global/types-kit'
-import type { HypernativeEligibility } from '@/features/hypernative/hooks/useIsHypernativeEligible'
+import type { HypernativeEligibility } from '@/features/hypernative'
 
 jest.mock('@safe-global/utils/features/safe-shield/hooks', () => ({
   useThreatAnalysis: jest.fn(),
@@ -49,7 +49,7 @@ const buildEligibility = (overrides: Partial<HypernativeEligibility> = {}): Hype
 
 const mockUseIsHypernativeEligible = jest.fn(() => buildEligibility())
 
-jest.mock('@/features/hypernative/hooks/useIsHypernativeEligible', () => ({
+jest.mock('@/features/hypernative', () => ({
   useIsHypernativeEligible: () => mockUseIsHypernativeEligible(),
 }))
