+++ apps/web/src/features/safe-shield/hooks/__tests__/useThreatAnalysis.test.tsx
@@ -3,7 +3,7 @@ import { useThreatAnalysis } from '../useThreatAnalysis'
 import { Severity, StatusGroup, ThreatStatus } from '@safe-global/utils/features/safe-shield/types'
 import type { SafeTransaction } from '@safe-global/types-kit'
 import { Safe__factory } from '@safe-global/utils/types/contracts'
-import type { HypernativeEligibility } from '@/features/hypernative/hooks/useIsHypernativeEligible'
+import type { HypernativeEligibility } from '@/features/hypernative'
 
 jest.mock('@safe-global/utils/features/safe-shield/hooks', () => ({
   useThreatAnalysis: jest.fn(),
@@ -50,7 +50,7 @@ const buildEligibility = (overrides: Partial<HypernativeEligibility> = {}): Hype
 
 const mockUseIsHypernativeEligible = jest.fn(() => buildEligibility())
 
-jest.mock('@/features/hypernative/hooks/useIsHypernativeEligible', () => ({
+jest.mock('@/features/hypernative', () => ({
   useIsHypernativeEligible: () => mockUseIsHypernativeEligible(),
 }))
