+++ apps/web/src/pages/transactions/history.tsx
@@ -14,9 +14,7 @@ import { BRAND_NAME } from '@/config/constants'
 import CsvTxExportButton from '@/components/transactions/CsvTxExportButton'
 import { useHasFeature } from '@/hooks/useChains'
 import { FEATURES } from '@safe-global/utils/utils/chains'
-import { useBannerVisibility } from '@/features/hypernative/hooks'
-import { BannerType } from '@/features/hypernative/hooks/useBannerStorage'
-import { HnBannerForHistory } from '@/features/hypernative/components/HnBanner'
+import { useBannerVisibility, BannerType, HnBannerForHistory } from '@/features/hypernative'
 
 const History: NextPage = () => {
   const [filter] = useTxFilter()
