+++ specs/001-migrate-hypernative/plan.md
@@ -1,0 +1,96 @@
+# Implementation Plan: Migrate Hypernative to Feature Architecture
+
+**Branch**: `001-migrate-hypernative` | **Date**: 2026-01-26 | **Spec**: [spec.md](./spec.md)
+**Input**: Feature specification from `/specs/001-migrate-hypernative/spec.md`
+
+## Summary
+
+Migrate the existing hypernative feature to comply with the feature architecture pattern defined in `apps/web/docs/feature-architecture.md`. This involves creating a main barrel file (`index.ts`) with lazy-loaded exports, updating ~25 external consumer files to import from the barrel, renaming the feature flag hook to `useIsHypernativeEnabled`, refactoring the OAuth callback page to a thin wrapper, and configuring ESLint rules to enforce import boundaries.
+
+## Technical Context
+
+**Language/Version**: TypeScript 5.x with Next.js 14.x
+**Primary Dependencies**: Next.js (dynamic imports for lazy loading), ESLint (eslint-plugin-boundaries, no-restricted-imports rule), Redux Toolkit (store slices)
+**Storage**: N/A (code migration only, no data storage changes)
+**Testing**: Jest with React Testing Library (existing test suite)
+**Target Platform**: Web (Next.js application in apps/web)
+**Project Type**: web (monorepo structure)
+**Performance Goals**: Bundle size must not increase; lazy-loading of feature components must be preserved
+**Constraints**: No functional regressions; all existing tests must pass; ESLint must pass
+**Scale/Scope**: ~25 external consumer files, ~50+ import statements to migrate, 1 hook rename, 1 page refactor
+
+## Constitution Check
+
+_GATE: Must pass before Phase 0 research. Re-check after Phase 1 design._
+
+| Principle                       | Status | Notes                                                                                                                   |
+| ------------------------------- | ------ | ----------------------------------------------------------------------------------------------------------------------- |
+| I. Type Safety                  | PASS   | No new code logic; existing types preserved. No `any` usage introduced.                                                 |
+| II. Branch Protection           | PASS   | Working on feature branch `001-migrate-hypernative`. Will run quality gates before commit.                              |
+| III. Cross-Platform Consistency | PASS   | Feature is web-only (`apps/web`); no changes to shared packages or mobile code.                                         |
+| IV. Testing Discipline          | PASS   | Existing tests updated for new import paths and hook name. No new MSW mocks needed.                                     |
+| V. Feature Organization         | PASS   | This migration specifically implements feature organization principles (barrel file, feature flags, import boundaries). |
+| VI. Theme System                | N/A    | No styling changes involved.                                                                                            |
+
+**Gate Result**: All applicable principles pass. Proceeding to Phase 0.
+
+## Project Structure
+
+### Documentation (this feature)
+
+```text
+specs/001-migrate-hypernative/
+├── spec.md              # Feature specification (complete)
+├── plan.md              # This file
+├── research.md          # Phase 0 output
+├── data-model.md        # Phase 1 output (minimal - code structure only)
+├── quickstart.md        # Phase 1 output
+├── contracts/           # Phase 1 output (N/A for this migration)
+├── checklists/          # Quality checklists
+│   └── requirements.md  # Spec quality checklist (complete)
+└── tasks.md             # Phase 2 output (generated by /speckit.tasks)
+```
+
+### Source Code (repository root)
+
+```text
+apps/web/src/
+├── features/
+│   └── hypernative/
+│       ├── index.ts                    # NEW: Main barrel file (public API)
+│       ├── components/
+│       │   ├── OAuthCallbackHandler/   # NEW: Moved from pages
+│       │   │   └── index.tsx
+│       │   └── [existing components]/
+│       ├── hooks/
+│       │   ├── index.ts                # Existing sub-barrel (internal use)
+│       │   ├── useIsHypernativeEnabled.ts  # RENAMED from useIsHypernativeFeature
+│       │   └── [existing hooks]/
+│       ├── services/
+│       ├── store/
+│       │   └── index.ts                # Existing sub-barrel (internal use)
+│       ├── config/
+│       ├── utils/
+│       └── constants.ts
+├── pages/
+│   └── hypernative/
+│       └── oauth-callback.tsx          # MODIFIED: Thin wrapper only
+├── components/                         # MODIFIED: Update imports
+├── store/
+│   └── slices.ts                       # MODIFIED: Update imports
+└── [other external consumers]/         # MODIFIED: Update imports
+
+# ESLint configuration
+apps/web/
+└── eslint.config.mjs                   # MODIFIED: Add import boundary rules
+```
+
+**Structure Decision**: The hypernative feature already resides in `src/features/hypernative/`. The migration adds a main barrel file and refactors imports across the codebase. No new directories are created except for the `OAuthCallbackHandler` component to hold the extracted OAuth logic.
+
+## Complexity Tracking
+
+No constitution violations requiring justification. The migration is a straightforward refactoring that reduces complexity by:
+
+- Consolidating external imports to a single entry point
+- Enforcing encapsulation through ESLint rules
+- Following established feature architecture patterns
